zephyr: make PPC/TCPC support optional
The "ppc" and "tcpc" properties are tagged as optional by the
devicetree.  Update the PPC and TCPC shim drivers to support an empty
entry in the ppc_chips[] and tcpc_config[] arrays.

This is useful for boards that need to setup different USB-C chips at
runtime and ensures there is a valid entry to update.

BUG=none
TEST=zmake build -a
TEST=locally, comment out the "ppc" and "tcpc" property on skyrim and
validate the corresponding arrays are correct.

Change-Id: Ief1c3670d5ef9d8e07a599286e77fbb9e229335e