charger: Add a way to disable debugging at build time
Add a way to disable some debugging related to the chgstate command. It
is a small, but enough to reclaim 70 bytes for damu.

Use functions to avoid needing #ifdefs in the main body of the code.
Convert the value to a bool while we are here.

Disable this for damu at the same time, since this CL otherwise causes
it to fail. This to free up some flash space so I can continue
refactoring. Space with clang increases from 6 to 140 bytes, almost
enough for an SMS message.

BUG=b:218332694
TEST=CC=clang make BOARD=damu
  *** 140 bytes in flash and 10916 bytes in RAM still available on damu RO ****

Change-Id: I07d2453aa74ab53a29275bba76cd2e8003c5dc14

damu: Disable CONFIG_CHARGE_STATE_DEBUG

BUG=b:218332694
TEST= CC=clang make BOARD=damu -j50
See that it has more space

Change-Id: I46b2a3fc85cf0f29d5fc4c986816fae760b9fd43