Agah: Fast forward RSMRST to PCH
On urgent shutdown, it's required for PCH_RSMRST to drop while PCH's
voltages are still above 95%. This CL adds a dedicated interrupt handler
for SEQ_EC_RSMRST_ODL (GPIOE2) to reduce the latency.

GPIOE2 belongs to MIWU module 0 (because it's on table 'Input
Assignments for MIWU0') and is connected to WUI81. WUI81 is connected to
group 8 (WKINTH) according to 'Multi-Input Wake-Up, Generic Block
Diagram'. So, its priority can be set by NPCX_MIWU0_GROUP_H.

BUG=b:279918234
TEST=Build

Change-Id: I35e49cc68fae98d20021a27865af8f9e38ade8e9