WIP: NPCX flash bank size
Increase the flash bank size from 64KB to 256KB. In practice, we only
need one bank size that's half the size of the flash device.

Using a larger bank size speeds up crec_flash_get_protect() because it
iterates over flash memory in bank size chunks, checking each bank's
protection status. In the future, we may need to push this call down to
the `cros_flash_driver_api` so it can be implemented with the advantage
of device specific knowledge.

rex:
  517:waiting for EC to allow higher power draw         859,094 (7,437)

BUG=b:270658478

Change-Id: I6910b3e6d556cfae0b2c350bee45464b97a6b57f