motion_sense: Fix null pointer reference in gesture detection
The motion_sense_fifo_stage_data with NULL were removed in CL:4499747.
But we need to stage and commit the motion state changed to activity
sensor for body detection on the ap side. Revert the change and Add
null pointer check in motion_sense_fifo_stage_data and fifo_stage_unit
to fix it.

BUG=b:272655176
TEST=./twister -v -i -p native_posix -p unit_testing -s
drivers/drivers.body_detection
TEST=received User proximity: Near/Far event from powerd log

Change-Id: I1ef882f25e07fa17d879d8b0a4b9ffe211b071e7