board/hyperdebug: Gpio monitoring verifies that list is same
The HyperDebug board can watch for edges on a set of pins, like a logic
analyzer.  It will buffer such events, for later retrieval by
OpenTitanTool.  The current logic depends on the list of signals at the
time of retrieval is exactly identical to the list given at start of
monitoring.  This was almost verified, in that there were checks for
pins being left our, or new pins being added to the list, but no check
for the order of the given names.

This CL adds a check that the order is the same.

BUG=b:281857275
TEST=tast run - gscdevboard.Ti50EcReset
TEST=Manual invocations of opentitantool to trigger error

Change-Id: Ifcd9b96a319aff94b073ca758224045a5b1d67d4