body_detection: Add support for parsing per sensor parameters
Previously the body detection parameters were set in Kconfig.
This didn't work well with boards that can use alternate motion sensor
for this feature. In such case we might want to set different parameters
depending on the sensor used.
In order to fix that parse the configuration from motion sensor data.
If it's not set fallback to Kconfig parameters.

No functional change for existing boards intended.

BUG=b:262680246
TEST=make -j48 try_build_boards
TEST=emerge-skyrim chromeos-zephyr

Change-Id: I68894d11322661d9b5e2e73ffe827175647f5555