ppc: ktu1125: Fix register used for VCONN control
The register that was being used to enable/disable VCONN
(KTU1125_SET_SW_CFG) was incorrect. It should be KTU1125_CTRL_SW_CFG.
In addition, VCONN was being applied to wrong CC line.

KTU1125_CC1S_VCONN --> applies VCONN to CC1 when the polarity is CC2
KTU1125_CC2S_VCONN --> applies VCONN to CC2 when the polarity is CC1

BUG=b:267589112
TEST=Verfied using totalphase that VCONN is being enabled and the CC
voltage of the CC data line is correct.

Change-Id: I45c85800849dbabeac36d976525de7d058ad23f4