corsola: Fix reported USB-C ports for no DB design with DB detection on
The definition of `CORSOLA_DB_NONE` was settled to no DB in the project
in the previous CL "corsola: Update implementation for no daughter board
", and the USB-C counts is determined by dts settings in previous CL "
corosla: Improve USB-C port configuration". This introduces a reported
port number mismatch when the DB detection is enabled, and the desired
detection result is no DB in the design.

Extend the DB status enum with `CORSOLA_DB_NO_DETECTION` to properly
describe the different DB status, and fix the reported USB-C port logic.

BUG=b:282625153
TEST=`zmake build <all board> --clobber`

Change-Id: Idd55381759677272e0bfc7b270647d759642ce01