fuel gauge: Refactor battery detection
This CL makes APIs in battery_fuel_gauge module use get_batt_params() to
access battery info (instead of calling get_batt_params to get an index
to board_battery_info array (a.k.a. 'battery type').

This allows us to use the same code for battery-info-in-cbi module,
which doesn't use an array of battery info.

BUG=b:283131235
TEST=Agah. Battery charges and discharges.

Change-Id: I4cdf3adfedb989303bfa040a2ea54312f0f45556