fuel gauge: Refactor battery detection
Currently fuel gauge module saves a detected battery as an index to
struct board_batt_params array. This CL makes the module save a battery
as a pointer.

This allows us to use the same code for battery-info-in-cbi module,
which doesn't use an array of battery info.

There is no functionality change except battery detection will be done
only once in HOOK_INIT or get_batt_params, whichever is called first.

BUG=b:283131235
TEST=Agah. Battery charges and discharges.

Change-Id: I4cdf3adfedb989303bfa040a2ea54312f0f45556