sm5803: implement more test cases
This adds or extends coverage for these driver functions:
 * sm5803_post_init
 * sm5803_check_vbus_level (errors)
 * sm5803_enable_linear_charge
 * sm5803_{enable,disable}_low_power_mode
 * sm5803_restart_charging and VBAT overvoltage interrupt
 * AP throttling on charger overtemp

LPM-related functions are modified to be safe in case of communication
errors; as written before they could write unexpected values if the reads
failed but writes succeeded when doing read-modify-write operations on
those registers. They now abort immediately if a read error occurs.

BUG=b:242544165
TEST=./twister -cCi -T zephyr/tests/drivers -s drivers.sm5803

Change-Id: I59c26bed16d7c82f66f1202432228d238490c1b7