starmie: Add AC/DC feedback lcd pin
Himax IC needs to set different touch parameters
according to the charging state. So here set up a
HOOK_AC_CHANGE event.when AC is plugged in,
Set EC_GPM6 is set to high level and low when AC
power is removed,EC_GPM6 is set to low level.

BUG=b:282117960
TEST=zmake build starmie
The test results are consistent with the expected
behavior.

Change-Id: Ief963661540e23d67147a17f4ecee464824e7000