starmie: Add AC/DC feedback lcd pin
Himax IC needs to set different touch parameters
according to the charging state. So need EC to
feedback AC/DC status pin.

BUG=b:282117960
TEST=zmake build starmie
The test results are consistent with the expected
behavior.
When the AC is plugged in,set EC_GPM6 to high.
When the AC is removed.set EC_GPM6 to low.

Change-Id: Ief963661540e23d67147a17f4ecee464824e7000