rex: Configure USB DB GPIOs at runtime
Rex has multiple USB daughterboard options. 4 of the signals going to
the USB DBs have DB type dependent definitions, so we need to configure
the affected GPIOs at runtime to match the connected DB.

Affected GPIOs are marked "no-auto-init" in the device tree and
configured at boot once the DB type has been determined. If the DB type
is not recognized, the pins are hi-Z'd with a pullup (i.e. no-connect).

BUG=b:267589112,b:267589042
TEST=boots on SKU3

Change-Id: Ida23dc0603646aba0a516fd9fa17a33237dc3bc3