RT1718S: Pull BC 1.2 functionality into a proper driver
The BC 1.2 functionality for this chip is currently entrenched within
the TCPC driver for it.  However, a board may use the TCPC without also
using the BC 1.2 functions.  Pull these out into their own driver file.

Note that it's assumed any consumer of this chip will still be using the
TCPC functionality, and the BC 1.2 module depends upon that.

BUG=b:283134516
TEST=CQ+1

Change-Id: I2d7504377855fa7a59d890a2e5bfa62d52467ef4