PCHG: Disable charging info after battery is full
Some peripheral devices do not stop charging session even after the
battery is fully charged. This causes a charger chip to consume extra
power to keep updating the EC.

This CL allows PCHG to instruct a charger chip to stop sending charging
info after a battery is full. Note that it does not stop charging.

BUG=b:270295252
TEST=TBD

Change-Id: I041a9706f1fff1ec85312c177798477d118bacdc