program/rex: enable ebuild for ISH variant of EC with USE flag
This commit modifies BUILD.py script to enable ebuild using
USE="ish_fw". "ISH_ENABLE" is the new flag introduced in program Kconfig
file which can be used to disable the EC sensing capability and i2c.
This approach eliminates the need to create a new variant,
"rex-sans-sensors" here. Therefore, if EC is built with this flag,
the EC image in /build/rex/firmware/<project> will not have any sensor
functions.
Following is the emake script patch required to enable this patch.
https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4557266

BUG=b'280329972
TEST=ebuild tested using the USE='ish_fw' flag. The EC does not have
     sensor capabilities and does not support sensor commands.

Change-Id: Ib688ae607e82ba48b12db872f5d4f72a387747b6
Signed-off-by: Jay Patel <jay2.patel@intel.com>