Hades: Adjust sequencer timing on power-on
The Intel power sequencer device (U56, SLG4BD44540VTR) on the Hades
needs to have adjusted timing to correctly power down the rails for the
unexpected power removal case. Long term (next build) we'll order
differently pre-programmed part but short term, we need an EC I2C write
to adjust the glitch filtering on the PPVAR_SYS (VBAT_MON pin).

The SLG4BD44540 is powered with PP3300_Z1 and should hold the value as
long as the device is powered.

BUG=b:282181312
TEST=build

Change-Id: I9bef23fc5335c64e10777b6ba9eb8533abac9e7e