bintools/llvm: Set path to bintools for ARM architecture
Provide path to bintools for ARM architecture basing on CROSS_COMPILE
variable (like zephyr does for gnu bintools).

This patch adds definitions of the CMAKE_C_COMPILER_tool-name because
CMake is looking for bintools by adding a suffix to compiler binary
e.g for AR it would be armv7m-cros-eabi-clang-ar, which doesn't exist.

BUG=b:239712345
TEST=zmake build bloonchipper (with 'llvm' toolchain)

Change-Id: I2769eef4ec120322a3d551cbe95cdb0a4116127b