keyboard_8042: test: Reafctor kblog console command
The keyboard event logging console command relies on malloc being
enabled to create a buffer for storing KBD events. We don't typically
use heap memory, so to add tests and clean up some tech debt, this CL
adjusts the kblog command to use a static ring buffer. The inclusion of
this feature and size of buffer are controllable via Kconfig (this
console command is off by default). This CL also adds unit tests.

BUG=b:283267124
TEST=./twister and on-device test

Change-Id: I8a0fd8701926b27a12314526f686e835b79dedbc