core/cortex-m: fix the assignment to __data_lma_start
In CL:4333199, the __data_lma_start is assigned to the address after the
end of the VMA of the .rodata section. However, this assignment is
broken when the end of the .rodata is not 8-byte alignment because the
CLNAG build seems always has the LMA of the .data section starting
from the 8-byte alignemnt address.

VMA      LMA           Size Align Out     In      Symbol
...
10063d5c 10063d9c       0     1   __data_lma_start = .
...
...
200baa98 10063da0      488    8   .data
200baa98 10063da0        0    1           . = ALIGN ( 4 )
200baa98 10063da0        0    1           __data_start = .
...

This CL fixes this the issue by changing the assignment from
	__data_lma_start = .;
to
	__data_lma_start = LOADADDR(.data) - SIZEOF(.header);

BUG=b:271333155,b:274937192
BRANCH=none
TEST=pass "make buildall -j3"
TEST=EC firmware can boot up and console is reponsive on the npcx EVB
with npcx9_evb and helipilot build.

Change-Id: Ibb35e196adf74ba307754e9452b4ac8a932d54a9
Signed-off-by: Jun Lin <CHLin56@nuvoton.com>