it83xx: cec: Add CEC support for ITE
Add an ITE version of the CEC bitbang driver.

Note, the ITE EC has a CEC peripheral and support for this will be added
in a later CL. But it only supports one CEC port, so the bitbang driver
can be used to enable a second CEC port.

BUG=b:270507438
TEST=With following CL, use CEC to turn a TV off and on on dibbi.
LOW_COVERAGE_REASON=Tests will be added in b:287558802

Change-Id: Ib633d38db9ca32224802c1ae2eec0c8c314d84a7