zephyr: emul_pi3usb9201: only build by default when testing
The emul_pi3usb9201 Kconfig option has been changed to allow building on
non ztest images in http://crrev.com/c/4501039, but that caused to be
build by default on the non-test image, using extra flash space and
confusing the build time checking script.

Add a conditional to the default option so that it only gets built by
default when testing, but can still be enabled manually on a normal
image.

BUG=none
TEST=zmake build tentacruel; zmake build magikarp

Change-Id: Ic444abcde52c65d005d34e265f11b67919490212