ovis: initial ec: Remove support for charging/batt/bc1.2
Ovis doesn't have either a battery or charger ic. This CL removes
support for those features. In addition bc1.2 is not supported on the
usbc ports, but as it interacts with the charging module, that support
is removed as well.

Lastly, the LED devicetree nodes are modified to
remove support for led_3/led_4 signals which are not present on Ovis
and the led_policy.dtsi is not included as the policy needs to be
updated to not depend on charging state or battery charge levels.

BUG=b:272128257
TEST=zmake build ovis is successful
LOW_COVERAGE_REASON=early-phase platform, b/285159904

Change-Id: I83f70daacf87d7a12e22542aa5c467aaf29c508a