rex: Allow VCONN swap when GPIO_EN_Z1_RAILS is asserted
The change to PWR_ALL_SYS_PWRGD results in VCONN swap being allowed
too late in the boot process which still leads to unexpected Hard
Resets of the PD. Since we actually want to supply VCONN as early
as possible EN_Z1_RAILS is the gpio that enables pp5000 which is
the earliest that VCONN should be supplied.

BUG=b:274576896, b:280356243
TEST=zmake build rex. Had Intel Validation team verify that there
were no longer any hard resets occurring and devices were ennumerated
properly.

Change-Id: Iecff63293b97606fa992f6d9f99c8c4f51af8f3e