system: fix GET_VERSION Host Command
Clear the response buffer of the EC_CMD_GET_VERSION host command.

Version 1 of the command has cros_fwid_ro and cros_fwid_rw buffers that
are set if CONFIG_CROS_FWID_VERSION is enabled. Otherwise the buffers
with random data are sent.

It is also important for the security, because we are sending previously
used stack without clearing it.

BUG=b:265450629
TEST=disable CONFIG_CROS_FWID_VERSION, run 'ectool version' on DUT and
make sure there are no unexpected data

Change-Id: I6941b2716d057b8e241dca6365193c87a05fdb0b