zephyr: Cleanup EC ASSERT options
CONFIG_ASSERT must be enabled to use asserts in Zephyr. CONFIG_ASSERT
is disabled by default. So PLATFORM_EC_DEBUG_ASSERT isn't actually
enabling asserts.

Add select ASSERT to PLATFORM_EC_DEBUG_ASSERT so ASSERTS are enabled in
Zephyr when this flag is turned on. Default to n since this matches the
current state and many boards cannot fit with asserts on.

PLATFORM_EC_DEBUG_ASSERT_BRIEF should cause the assert condition to be
dropped in the assert output. This is equivalent to the Zephyr option
ASSERT_NO_COND_INFO, so select this option.

BUG=b:285008455
TEST=`crash assert` on skyrim

Change-Id: I83077d5be85308a36f79c1ae890e5c5c91fb7381