charger: Combine the #if macros
Move these move together to reduce the amount of code. We don't need to
specific about which CONFIG caused the problem, since a simple
inspection of the source code is enough to see that.

Make the other #ifdefs unconditional so we can avoid needing #ifdef
later in the source.

For now, leave uvp_throttle_start_time as conditionally declared.

BUG=b:218332694
TEST=make buildall

Change-Id: I2ee2e77263f610073ccc5bd8d97bab57ebf0b078