Drop NVMEM migration from Cr50
It has been four years since the new nvmem layout was introduced, the
chance of encountering a device with an old ChromeOS and Cr50 images
are pretty slim, and if such device comes along and tries to update,
it will have to update to an intermediate ChroemOS version, which will
include Cr50 update and NVMEM migration.

Let's drop the NVMEM migration code from the Cr50 image, this results
in flash space Savings of 1468 bytes.

Unit tests are being updated to remove the migration testing.

BUG=b:
TEST=manual

Change-Id: I500c70d51361cb13d703e994932cba097d64a3b6