zmake: Fix parallel execution for compare-builds
The executor wait operation was included inside the for loop that kicks
off the configure/build step at each checkout.

Move the executor wait outside the loop so the builds of both checkouts
run in parallel.

Time to compare builds of 7 boards with change: 2m38s
Time to compare builds of 7 boards without change: 2m54s

BUG=none
TEST=zmake compare-builds

Change-Id: I13c057a04e288808d3244f282ff24c14d2f34a44