build.mk: npcx: generate the firmware header for test images
In the NPCX chip, the build.mk overwrites the cmd_ec_elf_to_flat of the
Makefile.rules to call ECST tool to generate the firmware header for the
RO image. However, it is only generated when the target name is
ec.RO.flat. This CL changes the build.mk to generate the header also for
unit test images. (EX. aes.RO.flat)

BUG=b:285319390
BRANCH=none
TEST=pass "make buildall"
TEST= "BOARD=helipilot make test-aes"; program aes.RO.flat to the flash; make sure EC can boot up and the console is responsive;
pass "runtest aes"

Change-Id: I174c13f34fb914cbebe331fbba53f875907d780d
Signed-off-by: Jun Lin <CHLin56@nuvoton.com>