run_device_test: Suppress a consider-using-with warning
Suppress a consider-using-with warning in flash_and_run_test. The
warning is valid in that a file resource is being created and
used. However, in this case using a "with" block would require us to
duplicate the "run_test" call so it makes more sense to suppress the
warning rather than use a "with".

BUG=none
TEST=passes upload hooks

Change-Id: I9a4433f4b4dfc3cab835802158b8215335eb0c51