cr50: make sure system DRBG is initialized early in the boot
System DRBG is usually initialized when randoms are requested. In most
of cases this happens as one of the first things AP does.

However, certain operations assumes system DRBG is available - e.g.
pair-wise consistency test in ECC key gen. If this operation is
invoked before DRBG is initialized, it will lead to failure.

In this CL with trigger DRBG initialization as last stage of board init
sequence.

BUG=b:262324344
TEST=manual

Change-Id: Iea573fb1c6166a399c6641ede87608334bd7f8c2
Signed-off-by: Vadim Sukhomlinov <sukhomlinov@google.com>