util: check_zephyr_project_config: switch BOARD_DIR to native_posix
The script validates config option, and currently uses BOARD_DIR as a
wildcard to capture board options for all possible boards.

That was copied from the documentation generation code, where it made
sense, but it's unnecessary for the scope of this script, and has the
unwanted side effect of exposing it to any board Kconfig
misconfiguration, especially with option defined in modules we don't
use, for example this is now firing some false positive:

      warning: LV_Z_DOUBLE_VDB...
      warning: LV_Z_FULL_REFRESH...
      warning: LV_Z_VDB_ALIGN...
      warning: LV_Z_VBD_CUSTOM_SECTION...

Point BOARD_DIR to native_posix, just to give the script something to
work with. This should make it less susceptible to unrelated breakagaes.

BUG=none
TEST=util/check_zephyr_project_config.py -d zephyr/program/dev/prj.conf

Change-Id: I4b4bae2b3adf9dca0c82ad2ec0b416548a00665c