util: check_zephyr_project_config: skip symbols with "default y if ZTEST"
We recently started allowing some emulators to be build in non ZTEST
builds, those should not follow the usual pattern of "default y depends
on DT_HAS_..." but rather "default y in ZTEST" so that they don't get
accidentally built in production images. In that case though they have
to be enabled explicitly for dev platforms, which sauses the kconfig
check script to fail.

Add an explicit check for this condition so that the script does not
complain about those.

BUG=none
TEST=util/check_zephyr_project_config.py -d zephyr/program/dev/prj.conf

Change-Id: I05cf9ad5d95124271433d9d75f4e955cd1e416c1