rex: remove charger input current limit minimum value
Rex sets the minimum charger power with no battery to 30W, and the
USB-PD configuration means that input power with input current clamped
to 512 mA minimum can never exceed 10W: not enough power to keep the
system running.

Unset the input current minimum because it's currently insufficient to
keep a device alive with no battery and causes the system to exceed
pSnkStdby during VBUS voltage transitions when either the start or end
voltage is greater than 5V.

This also updates ovis, because ovis is a copy of rex.

BUG=b:163093572
TEST=builds

Change-Id: Ibe57829c94fc46f38999b4b894e47254c419d3a5