chip/stm32/usb: Fix the hammer cannot wakeup the host
The logic of enable/disable USB clock register were refactor in the
CL:3130733. When usb_suspend, the USB clock will be disable, which
leads to the hammer cannot wakeup the host. On the firmware-kukui
branch [1] that used to work fine, there was no enable/disable USB
clock logic in the clock_enable_module, so it do nothing there.
Maybe we can align the way it worked before.

[1] https://chromium.googlesource.com/chromiumos/platform/ec/+/refs
/heads/firmware-kukui-12573.B/chip/stm32/clock-stm32f0.c#420

BUG=b:284059785
TEST=powerd_dbus_suspend on the host, and wakeup the host by hammer

Change-Id: Ie600912aa0a86e2a9b2a03562314dc00e297ee28