flash_ec: Improve error message for npcx_monitor
When npcx_monitor.bin is not found, flash_ec simply states such.
This is unclear to users, particularly when only trying to load
pre-built ec binary images.  If the user specifies a pre-built image,
and npcx_monitor.bin is not found, then flash_ec now provides a more
descriptive error to the user on how to resolve the issue.

BUG=b:287304704
TEST=util/flash_ec --board hades --image /tmp/ec_hades_0613.bin

Change-Id: Iec75634978d679830f9d9b9cfc5b9514ea29f200