AMD Power: Add suspend exit on AP shutdown
If the AP shuts down while we're in S0ix, we currently have no path to
shut down the system and will remain in S0ix.  Add a check for the S5
rail from the AP being lost, and trigger a similar shutdown path that
would happen on power loss.

BUG=b:285596823
TEST=./twister -T ./zephyr/test/amd_power, on skyrim, run "hibernate"
from S0ix and ensure we power sequence down

Change-Id: I7bc961f1870f2abb157043dc406c3582d27e0870