i8042: Send pending scancode immediately after SETLEDS timer expires
When a host sends SETLEDS, EC waits for the 2nd byte and starts a timer.
When the timer expires, the EC gives up (close the SETLEDS session) and
sends a scancode.

This CL makes keyboard_protocol task adjust 'wait' length and read a
scancode from a queue so that a pending scancode will be sent to the
host immediately after the timer expires.

BUG=b:273636117
TEST=Vell. Platform 15511.0.0.

Change-Id: I54115ecfd0db6b25270eda72ba46a4ce50f1186d