cmake: Only add -mno-unaligned-access for ARM
The -mno-unaligned-access flag is for ARM architectures. We ought to
only include it when building for ARM.

Add guard that only includes the -mno-unaligned-access flag if ARM
architecture is used.

Note: CL:4621542 was an earlier attempt to guard the flag using
CONFIG_SOC_FAMILY_ARM. This attempt uses CONFIG_ARM.

BUG=b:268051452
TEST=CQ
TEST=Build rex and check -mno-unaligned-access flag was used
TEST=Build rex and use the compare builds script
TEST=Build WIP fwsdk twister

Change-Id: If95c19d5d974b8f69913562eb51bbc57a1657c7c