cec: Support receiving messages from multiple ports
Currently, CEC messages are passed to the AP in a cec_message MKBP
event. Since the max event size is 16 bytes, there's no space for a port
parameter. We could increase the event size, but this would increase
AP-EC traffic for all event types.

Instead, use a MKBP event EC_MKBP_CEC_HAVE_DATA to notify the AP that
data is ready, and add a read command for the AP to read the data.

To make ectool backwards compatible with old EC firmware, it supports
both methods of receiving messages. The kernel driver will also support
both.

To make the EC firmware backwards compatible with old kernel/ectool, if
a device only has one CEC port then it will continue to use cec_message.
This makes migration easier since it doesn't matter if the EC or kernel
changes land first.

BUG=b:270507438
TEST=`ectool cec read` still works on dibbi

Change-Id: I37590dd688ca4f994879ccefa40a4a81528c1b6c