cbi: flash: Set the CBI image size to 256 bytes
The CBI section size on EC flash is different from CBI image size. CBI
section size is dependent on erase block size and configured in binman
DT. The CBI image size is 256 bytes irrespective of the image location
EEPROM or EC flash. CBI data is making "bss" section overflow if it's
greater than 4096 bytes when loaded into RAM. So CBI image size cannot
be equal to it's section size in flash if it's greater than 4096 bytes.
To have uniformity it's set to 256 bytes similar to EEPROM CBI image.
Doing that will enable CBI driver to perform CBI I/O operation on first
256 bytes of CBI section on EC flash but flash driver will preserve all
of the CBI section on flash irrespective of image size.

BUG=b:267628657
TEST=Flashed on Rex

Change-Id: If72d76deab1015888c69228270b2d6e3a3800d46