tpm: rename reset_in_progress and use an enum
reset_in_progress isn't exactly tracking if there's a reset in progress.
It means the tpm could be resetting or cr50 could be waiting for AP
communication. It's used to debounce TPM_RST_L pulses until there's ap
communication, so cr50 doesn't reset the tpm back to back.

This change renames some things and starts using an enum instead of 0/1
for the value, so we can add more states later. It doesn't change any
behavior.

BUG=b:233319621
TEST=run firmware_Cr50DeviceState

Change-Id: I3931086bb6661f3a2854aefe4705aaaf08be2a9f