TCPMv2: PRL: Check PRL_FLAGS_TX_COMPLETE for extended messages
Currently when sending an extended message, the PRL layer will notify
the PE that the message send completed before receiving confirmation
from the TCPC.

This CL updates the PRL layer to check for PRL_FLAGS_TX_COMPLETE before
the TCH_Message_Sent state transition when sending an extended message.
It creates consitency with the TCH_Wait_For_Transmission_Complete to
TCH_Message_Sent transition for non-extended messages and prevents the
PRL layer from notifying the PE that a message transmission completed
before confirmation from the TCPC.

BUG=b:283737121
TEST=make try_build_boards, make runhosttest, manually checked device
enumeration with a dock which requests an extended message (Status).

Change-Id: Ia639f55cb824fc65adf8df3bab9185b1f291387d