zephyr/linker: Add NOLOAD attribute to .end_of_ram_info section
The section is used to provide start, end and size of unused RAM and
make some checks. Currently, this section has LOAD flag which means that
it's part of image layout. Attempt to generate binary file with
--gap-fill results in creating a very big file, because there is a gap
between the end of .end_of_ram_info section (0x0 address) and beginning
of flash (eg. 0x8000000 for STM32F4).

This problem doesn't appear when linking using GNU LD, because it
removes the section in garbage collect process. However, LLVM LLD
doesn't garbage collect sections that define used symbols, so the
section is present in final ELF image.

There is no need to load this section (it has 0 size), so we can safely
add NOLOAD attribute to the section.

BUG=b:239712345
TEST=zmake build bloonchipper

Change-Id: Id22f33134e64d44fa6918d3d525644b4e5ff771b