zephyr: Move ram size check out of linker to Python script
This is a downstreaming unblocker. An upstream change modified the order
in which bits of the linker script are assembled. Long story short, a
linker snippet added by the EC codebase to enforce that enough free RAM
remains for crash data (`end-of-ram.ld`) now malfunctions due to
`_image_ram_end` not being defined yet. I could not find a practical way
to get this inserted into the correct spot. In the interest of reducing
complexity in the linker scripts, I re-implemented the same checks in a
Python script that examines the outputted ELF file post-build.

Also, an additional check involving the minimum shared memory size was
moved to a build-time assertion in `common/shared_mem.c`

Two boards actually currently fail the size check (Gothrax and Nereid).
This was not apparent earlier because the ITXXX2 linker script uses the
variable `RAM_BASE` instead of the expected `RAM_ADDR` (The check was
being skipped if `RAM_ADDR` was undefined). Bugs filed for these two
boards to sort out later.

BUG=None
TEST=zmake build -a against the problem Zephyr commit

Change-Id: Ied36534d74454110ad81e6966f9d1a3e4e197c88