cec/bitbang: Add a port parameter to functions
Add a port parameter to all functions of the CEC bitbang driver. This
will allow the driver to support multiple ports in the future.

For now, ITE and Nuvoton both only support one instance of the driver,
so a fixed port is used. If support for multiple ports is added in the
future, the port should be detected based on the timer/GPIO where the
interrupt occurred.

BUG=b:270507438
TEST=CEC works on dibbi

Change-Id: I690643dc6afa21a7fa3887091f8be43b02305254