cec/bitbang: Configure GPIOs in cec_config instead of using macros
Add a field to cec_config_t for storing driver-specific configuration
data. For the bitbang driver, use this to configure the input, output
and pull-up GPIOs. This replaces the CEC_GPIO_* macros currently used,
and will make it easier to run more than one instance of the bitbang
driver in the future.

BUG=b:270507438
TEST=CEC works on dibbi

Change-Id: I0139dfc3913de2a056b93f983d3c373205aea10d