rng_benchmark: Reduce test_trng_rand_bytes_toggle iterations
This test is very time consuming, especially on helipilot where it
takes roughly 1 second to init / exit RNG. On Helipilot this results
in run_device_tests timing out. Constantly cycling the RNG isn't
really a typical use case but the test still has some value in
ensuring the results are changing. The number of iterations isn't
really critical here so reduce the number to 10 to prevent timeouts.

BUG=b:286232426
TEST=run_device_tests.py --board helipilot -t rng_benchmark => PASS

Change-Id: Ie00a325a0be7dad636dc768d7eaf22d19051482b