bazel: Add flash_ec functionality
This generates a run target for each EC target.  The run target has
the corresponding name ":flash_ec_{board}", and calls flash_ec for
that board.

The dependency is automatically added on the build, so builds will
happen automatically as required before the flash.

Note that since we don't support Legacy EC yet, this does not yet
support Legacy EC flashing.

BUG=b:286125027
TEST=bazel run :flash_ec_brya

Change-Id: I5f692d699ad1cfe28ebc3fbb8f8a6dd54b69f04b