zephyr: keyboard: check for concurrent build of upstream and internal npcx keyboard driver
Add a build assert to validate that the internal NPCX raw keyboard
driver is not built at the same time as the upstream one, since the two
drivers use the same underlying hardware.

Change the default for the upstream symbol so that it's not enabled
automatically when the input subsystem is enabled and also make sure
that keyboard raw driver depends on the keyboard raw symbol.

BUG=b:274443029
TEST=zmake compare-builds -a

Change-Id: Iff33b592b4cb7e7edec1e810b306a0c1f0a1fa06